+++ lib-mal/CHANGELOG.md
@@ -8,3 +8,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Oauth2 authorization 
+- `Get anime list` API funciton
+- `Get anime details` API function